refactor: moving the hdinsight client
